+++ projects/cq-angular-editable-components/src/lib/layout/aem-component.directive.ts
@@ -15,12 +15,9 @@
  * from Adobe Systems Incorporated.
  */
 
-import { Directive, Input, Output, EventEmitter, Renderer2, NgZone, ChangeDetectorRef, ViewContainerRef, ComponentFactoryResolver, ApplicationRef, ComponentRef } from '@angular/core';
+import { Directive, Input, Renderer2, NgZone, ViewContainerRef, ComponentFactoryResolver, ComponentRef, AfterViewInit } from '@angular/core';
 
 import { ComponentMapping } from "./component-mapping";
-
-import { PageModelManager } from '@adobe/cq-spa-page-model-manager';
-
 import { Constants } from "./constants";
 import { Utils } from "./utils";
 
@@ -31,14 +28,14 @@ const PLACEHOLDER_CLASS_NAME = 'cq-placeholder';
   selector: '[aemComponent]'
 })
 
-export class AEMComponentDirective {
+export class AEMComponentDirective implements AfterViewInit {
   private _component:ComponentRef<any>;
-  private _oldClasses:string;
 
-  @Input() cqModel:any;
-  @Input() path:string;
-  @Input() pagePath:string;
-  @Input() modelName:string;
+  @Input() cqItem:any;
+  @Input() cqPath:string;
+  @Input() itemName:string;
+  @Input() itemAttrs: object;
+
   @Input() aemComponent;
 
   constructor(
@@ -56,44 +53,8 @@ export class AEMComponentDirective {
    * Returns the type of the cqModel if exists.
    */
   get type() {
-    return this.cqModel && this.cqModel[":type"];
+    return this.cqItem && this.cqItem[Constants.TYPE_PROP];
   }
-
-  /**
-   * Returns the column classes of the cqModel
-   */
-  get columnClasses() {
-    return this.cqModel && (this.cqModel.columnClassNames || '');
-  }
-
-  /**
-   * Updates the cqModel
-   */
-  private updateCqModel() {
-    let self = this;
-    // Fetching the latest data for the item at the given path
-    this.getCqModel().then(model => {
-        this.ngZone.run(() => {
-          model[Constants.DATA_PATH_PROP] = this.path;
-          this.cqModel = model;
-          this.updateComponentData();
-          this.setupElement();
-          let editConfig = ComponentMapping.getEditConfig(this.type);
-          if (editConfig) {
-            this.setupPlaceholder(editConfig);
-          }
-        });
-    });
-  }
-
-  /**
-   * Returns the Cq Model
-   *
-   */
-  private getCqModel() {
-      return PageModelManager.getData({pagePath: this.pagePath, dataPath: this.path});
-  }
-
   /**
    * Renders a component dynamically based on the component definition
    *
@@ -104,15 +65,7 @@ export class AEMComponentDirective {
       const factory = this.factoryResolver.resolveComponentFactory(componentDefinition);
       this.viewContainer.clear();
       this._component = this.viewContainer.createComponent(factory);
-
       this.updateComponentData();
-      this.setupElement();
-      let editConfig = ComponentMapping.getEditConfig(this.type);
-      if (editConfig && Utils.isInEditor) {
-        this.setupPlaceholder(editConfig);
-      }
-      PageModelManager.removeListener({pagePath: this.pagePath, dataPath: this.path, callback: this.updateCqModel.bind(this) });
-      PageModelManager.addListener({pagePath: this.pagePath, dataPath: this.path, callback: this.updateCqModel.bind(this) });
     }
   }
 
@@ -120,76 +73,55 @@ export class AEMComponentDirective {
    * Updates the data of the component based the data of the directive
    */
   private updateComponentData() {
-    let keys = Object.getOwnPropertyNames(this.cqModel);
+    let keys = Object.getOwnPropertyNames(this.cqItem);
+
     keys.forEach((key) => {
-      if (key !== ":type") {
-        let propKey = key.startsWith(":") ? key.substr(1) : key;
-        // we need to wrap so we get the underlying data.
-        // this way data-binding will work.
+        let propKey = key;
+
+        if (propKey.startsWith(":")) {
+            // Transformation of internal properties namespaced with [:] to [cq]
+            // :myProperty => cqMyProperty
+            let tempKey = propKey.substr(1);
+            propKey = "cq" + tempKey.substr(0, 1).toUpperCase() + tempKey.substr(1);
+        }
+
+        const descriptor = Object.getOwnPropertyDescriptor(this._component.instance, propKey);
+
         Object.defineProperty(this._component.instance, propKey, {
-          get: () => { return this.cqModel[key]; },
-          set: (value) => { this.cqModel[key] = value}
+          get: () => { return this.cqItem[key]; },
+          set: (value) => { this.cqItem[key] = value}
         });
-      }
     });
 
-    this._component.instance.path = this.path;
-    this._component.instance.pagePath = this.pagePath;
-    this._component.instance.modelName = this.modelName;
+    this._component.instance.cqPath = this.cqPath;
+    this._component.instance.itemName = this.itemName;
   }
 
   /**
-   * Setups the DOM element, setting the classes and attributes needed for the AEM editor.
+   * Adds the specified item attributes to the element
    */
-  private setupElement()  {
-    if (this._oldClasses) {
-      let oldClasses = this._oldClasses.split(' ');
-      oldClasses.forEach((columnClass) => {
-        this.renderer.removeClass(this._component.location.nativeElement, columnClass);
+  private setupItemAttrs() {
+    if (this.itemAttrs) {
+      let keys = Object.getOwnPropertyNames(this.itemAttrs);
+
+      keys.forEach((key) => {
+        if (key === "class") {
+          let classes = this.itemAttrs[key].split(' ');
+          classes.forEach((itemClass) => {
+            this.renderer.addClass(this._component.location.nativeElement, itemClass);
+          });
+        } else {
+          this.renderer.setAttribute(this._component.location.nativeElement, key , this.itemAttrs[key])
+        }
       });
     }
-
-    this._oldClasses = this.columnClasses;
-    // Manually add the classes
-    if (this.columnClasses) {
-      let classes = this.columnClasses.split(' ');
-      classes.forEach((columnClass) => {
-        this.renderer.addClass(this._component.location.nativeElement, columnClass);
-      });
-    }
-
-    if (this.path) {
-      this.renderer.setAttribute(this._component.location.nativeElement, "data-cq-data-path" ,this.path);
-    }
-
   }
 
-  /**
-   * Setups the placeholder of needed for the AEM editor
-   *
-   * @param editConfig - the editConfig, which will dictate the classes to be added on.
-   */
-  private setupPlaceholder(editConfig) {
-    if (this.usePlaceholder(editConfig)) {
-        this.renderer.addClass(this._component.location.nativeElement, PLACEHOLDER_CLASS_NAME);
-        this.renderer.setAttribute(this._component.location.nativeElement, "data-emptytext", editConfig.emptyLabel);
-    } else {
-        this.renderer.removeClass(this._component.location.nativeElement, PLACEHOLDER_CLASS_NAME);
-        this.renderer.removeAttribute(this._component.location.nativeElement, "data-emptytext");
-    }
-  }
-
-  /**
-   * Determines if the placeholder should e displayed.
-   *
-   * @param editConfig - the edit config of the directive
-   */
-  private usePlaceholder(editConfig) {
-    return editConfig.isEmpty && typeof editConfig.isEmpty === "function" && editConfig.isEmpty(this.cqModel);
+  ngAfterViewInit() {
+    this.setupItemAttrs();
   }
 
   ngOnDestroy() {
-    PageModelManager.removeListener({pagePath: this.pagePath, dataPath: this.path, callback: this.updateCqModel.bind(this) });
     this._component && this._component.destroy();
   }