+++ projects/cq-angular-editable-components/src/lib/layout/aem-container/aem-container.component.ts
@@ -16,23 +16,34 @@
  */
 
 import { Component, Input } from '@angular/core';
-import { Constants } from '../constants';
+import { Constants } from '../constants'
+import { Utils } from "../utils";
+
+const PLACEHOLDER_CLASS_NAMES = Constants.NEW_SECTION_CLASS_NAMES;
+const PLACEHOLDER_ITEM_NAME = '*';
+const CONTAINER_CLASS_NAMES = 'aem-container';
 
 @Component({
   selector: 'aem-container',
   host: {
-      '[attr.data-cq-data-path]':'path'
+      '[class]': 'hostClasses',
+      '[attr.data-cq-data-path]':'cqPath'
   },
   templateUrl: './aem-container.component.html'
 })
 export class AEMContainerComponent {
-  @Input() items;
-  @Input() itemsOrder;
-  @Input() path:string = '';
-  @Input() pagePath:string = '';
+  @Input() cqItems;
+  @Input() cqItemsOrder;
+  @Input() cqPath:string = '';
   @Input() modelName:string = '';
+  @Input() classNames: string;
 
-  constructor() {}
+  /**
+   * Returns weather of not we are in the editor
+   */
+  get isInEditMode() {
+    return Utils.isInEditor();
+  }
 
   /**
    * Returns the aggregated path of this container path and the provided path
@@ -40,7 +51,7 @@ export class AEMContainerComponent {
    * @param path - the provided path to aggregate with the container path
    */
   getDataPath(path) {
-    return this.path ? this.path + "/" + path : path;
+    return this.cqPath ? this.cqPath + "/" + path : path;
   }
 
   /**
@@ -49,7 +60,32 @@ export class AEMContainerComponent {
    * @param itemKey - the itemKey to look for in the items.
    */
   getItem(itemKey) {
-    return this.items && this.items[itemKey];
+    return this.cqItems && this.cqItems[itemKey];
+  }
+
+  /**
+   * Returns the class names of the container based on the data from the cqModel
+   */
+  getHostClassNames() {
+    return CONTAINER_CLASS_NAMES;
+  }
+
+  get hostClasses () {
+    return this.getHostClassNames();
+  }
+
+  /**
+   * Returns the placeholder classes
+   */
+  getPlaceholderClassNames() {
+    return PLACEHOLDER_CLASS_NAMES;
+  }
+
+  /**
+   * Returns the placeholder path
+   */
+  get placeholderPath() {
+    return this.cqPath && this.cqPath + '/' + PLACEHOLDER_ITEM_NAME;
   }
 }