+++ projects/cq-angular-editable-components/src/lib/layout/aem-container/aem-container.component.spec.ts
@@ -21,7 +21,9 @@ import { BrowserDynamicTestingModule } from '@angular/platform-browser-dynamic/t
 
 import { AEMContainerComponent } from './aem-container.component';
 import { AEMComponentDirective } from '../aem-component.directive';
+import { AEMModelProviderComponent } from '../aem-model-provider/aem-model-provider.component';
 import { AEMResponsiveGridComponent } from "../aem-responsivegrid/aem-responsivegrid.component";
+import { ModelManager, ModelStore } from "@adobe/cq-spa-page-model-manager";
 import { Component1 } from "../../test/test-comp1.component";
 import { Component2 } from "../../test/test-comp2.component";
 import { Component3 } from "../../test/test-comp3.component";
@@ -34,15 +36,17 @@ describe('AEMContainerComponent', () => {
   let fixture: ComponentFixture<AEMContainerComponent>;
 
   beforeEach(async(() => {
+    let modelStore = new ModelStore('rootPath', require("../../test/data/layout.json") );
+    ModelManager.initialize({ path: 'rootPath', modelStore });
     TestBed.configureTestingModule({
-      declarations: [ AEMContainerComponent, AEMComponentDirective,
+      declarations: [ AEMContainerComponent, AEMComponentDirective, AEMModelProviderComponent,
       Component1,
       Component2,
       Component3,
       AEMResponsiveGridComponent ]
     }).overrideModule(BrowserDynamicTestingModule, {
       set: {
-        entryComponents: [Component1, Component2, Component3, AEMResponsiveGridComponent]
+        entryComponents: [Component1, Component2, Component3, AEMResponsiveGridComponent, ]
       }
     })
     .compileComponents();
@@ -71,54 +75,37 @@ describe('AEMContainerComponent', () => {
     fixture.detectChanges();
     let element = fixture.nativeElement;
     element = checkComponent(element.firstElementChild,
-      "aem-responsivegrid", "root", "aem-container");
+      "aem-responsivegrid", "root", "aem-container aem-Grid/root");
 
     element = checkComponent(element.firstElementChild.firstElementChild, "aem-responsivegrid",
-      "root/responsivegrid", "aem-container aem-GridColumn/root/responsivegrid");
+      "root/responsivegrid", "aem-container aem-Grid/root/responsivegrid");
 
-    element = checkComponent(element.firstElementChild.firstElementChild, "test-comp1",
+    element = checkComponent(element.firstElementChild, "div",
       "root/responsivegrid/component1",
-      "aem-GridColumn/root/responsivegrid/component1", "Component1");
+      "aem-GridColumn/root/responsivegrid/component1");
+    expect(element.firstElementChild.matches(`test-comp1[data-title="Component1"]`)).toBeTruthy();
 
-    element = checkComponent(element.nextElementSibling, "test-comp3",
+    element = checkComponent(element.nextElementSibling, "div",
       "root/responsivegrid/component3",
-      "aem-GridColumn/root/responsivegrid/component3", "Component3");
+      "aem-GridColumn/root/responsivegrid/component3");
+    expect(element.firstElementChild.matches(`test-comp3[data-title="Component3"]`)).toBeTruthy();
 
-    element = checkComponent(element.nextElementSibling, "test-comp1",
+    element = checkComponent(element.nextElementSibling, "div",
       "root/responsivegrid/component5",
-      "aem-GridColumn/root/responsivegrid/component5", "Component5");
+      "aem-GridColumn/root/responsivegrid/component5");
+    expect(element.firstElementChild.matches(`test-comp1[data-title="Component5"]`)).toBeTruthy();
 
-    element = checkComponent(element.nextElementSibling, "test-comp2",
+    element = checkComponent(element.nextElementSibling, "div",
       "root/responsivegrid/component2",
-      "aem-GridColumn/root/responsivegrid/component2", "Component2");
+      "aem-GridColumn/root/responsivegrid/component2");
+    expect(element.firstElementChild.matches(`test-comp2[data-title="Component2"]`)).toBeTruthy();
 
-    element = checkComponent(element.nextElementSibling, "test-comp2",
+    element = checkComponent(element.nextElementSibling, "div",
       "root/responsivegrid/component4",
-      "aem-GridColumn/root/responsivegrid/component4", "Component4");
+      "aem-GridColumn/root/responsivegrid/component4");
+    expect(element.firstElementChild.matches(`test-comp2[data-title="Component4"]`)).toBeTruthy();
+
     expect(component).toBeTruthy();
-  });
-
-  it("updates the title", () => {
-    let layout = require("../../test/data/layout.json");
-    component.cqItems = layout[Constants.ITEMS_PROP];
-    component.cqItemsOrder = layout[Constants.ITEMS_ORDER_PROP];
-    component.cqPath = layout[Constants.PATH_PROP];
-
-    fixture.detectChanges();
-    let element = fixture.nativeElement.querySelector('test-comp1[data-cq-data-path="root/responsivegrid/component1"]');
-    expect(element).toBeDefined();
-    expect(element.getAttribute("data-title")).toBe("Component1");
-    let component1 = layout[Constants.ITEMS_PROP]["root"][Constants.ITEMS_PROP]["responsivegrid"][Constants.ITEMS_PROP]["component1"];
-
-    let changedTitle = "Changed Title";
-    component1.title = changedTitle;
-    fixture.detectChanges();
-    expect(element.getAttribute("data-title")).toBe(changedTitle);
-
-    changedTitle = "Some other one";
-    component1.title = changedTitle;
-    fixture.detectChanges();
-    expect(element.getAttribute("data-title")).toBe(changedTitle);
   });
 
   it('should create placeholder', () => {