+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### 1.0.0 - 28 September 2018
+### 1.0.2 - 28 September 2018
 
 * Adapting to the new PageModelManager API
 * Refactoring the Container to be independent of the ModelManager. This is achieved by refactoring the aem-directive to only be in charge of creating the dynamic components. This way consumers can add their own logic to update the model, such as from a store.